% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2alpha.R
\name{col2alpha}
\alias{col2alpha}
\title{Convert color to transparent color}
\usage{
col2alpha(col, alpha = 0.5)
}
\arguments{
\item{col}{Character, the color}

\item{alpha}{Length 1 numeric, 0 to 1 (the transparency level)}
}
\description{
Convert color to transparent color
}
