% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelLines.R
\name{labelLines}
\alias{labelLines}
\title{Label the line positions on a plot}
\usage{
labelLines(outer = FALSE, alpha = 1)
}
\arguments{
\item{outer}{Logical of length 1, should the outer lines be drawn?}

\item{alpha}{Length 1 numeric, 0 to 1 (the transparency level)}
}
