% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line2user.R
\name{line2user}
\alias{line2user}
\title{Convert the line location to user coordinates}
\usage{
line2user(line, side, outer = FALSE)
}
\arguments{
\item{line}{Integer, the line}

\item{side}{Integer, the side}

\item{outer}{Logical, should the outer margin be used?}
}
\description{
Convert the line location to user coordinates
}
