% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.R
\name{primers}
\alias{primers}
\title{Messy primers data}
\description{
A 1-column data.frame with primer strings
}
\details{
The data is intentionally "messy" to demonstrate issues with
file loading, and how to clean data.
}
