% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRClass.R
\name{updateRClass}
\alias{updateRClass}
\title{Pull & build/rebuild package from GitHub}
\usage{
updateRClass(ref = "master")
}
\arguments{
\item{ref}{Passed to \code{install_github 'ref'}}
}
\description{
Pull & build/rebuild package from GitHub
}
