% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.R
\name{yeast}
\alias{yeast}
\title{Time-series data for three yeast strains}
\description{
Time-series data for three yeast strains
}
\details{
In wide format with strain in a different column. 
Provided by Amy Pomeroy
}
