% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.R
\name{yeast_long}
\alias{yeast_long}
\title{Long-format time-series data for three yeast strains}
\description{
Long-format time-series data for three yeast strains
}
\details{
In long format with strain given in the 'Strain' column. 
Provided by Amy Pomeroy
}
